[#29] refactor: Auth 관련 메서드 Pinia 사용하도록 수정
- close #29
